Move `dat_to_csv` to `external.py`
