+++ deepxde/utils/external.py
@@ -1,5 +1,6 @@
 """External utilities."""
 
+import csv
 from multiprocessing import Pool
 
 import numpy as np
@@ -126,3 +127,23 @@ def uniformly_continuous_delta(X, Y, eps):
             if np.max(dY[dX <= delta]) >= eps:
                 return delta - dx / 2
             delta += dx
+
+
+def dat_to_csv(dat_file_path, csv_file_path, columns):
+    """Converts a dat file to CSV format and saves it.
+
+    Args:
+        dat_file_path (string): Path of the dat file.
+        csv_file_path (string): Desired path of the CSV file.
+        columns (list): Column names to be added in the CSV file.
+    """
+    with open(dat_file_path, "r", encoding="utf-8") as dat_file, open(
+        csv_file_path, "w", encoding="utf-8", newline=""
+    ) as csv_file:
+        csv_writer = csv.writer(csv_file)
+        csv_writer.writerow(columns)
+        for line in dat_file:
+            if "#" in line:
+                continue
+            row = [field.strip() for field in line.split(" ")]
+            csv_writer.writerow(row)
